QNX: Make window groups work reliably
Window group related code worked previously only for the 'root'
window and other windows that had no parents (which then joined
the window group of the root).

Introduced a new function joinGroup that makes the group related
calls, this can both join and leave groups.

Default group joining is performed in setVisible now rather than
setScreen.  The advantage of setVisible is that one can rely on it
getting called at some point but it is also possible to create and
interact with a QQnxWindow before it is called.  This allows
the default group handling to be overridden via the native
interface.  If the group is set (or explicitly not set) then the
default handling is not done.

joinGroup is also called from setParent so that if the parent window
is changed the window joins the parent group.

Change-Id: I03e3801ecb4e4cf901c02e45f00472a87692c2ec
Reviewed-by: Rafael Roquetto <rafael.roquetto@kdab.com>
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>