Mac: Implement unified toolbar look for QTabBar
Extend the unified title and toolbar gradient to 
tabs in document mode that are adjacent unified
tool bars.

Change the updateMacBorderMetrics() function to
register the tab bar geometry and visibility status
with the Cocoa platform plugin. The Cocoa platform
plugin will then merge this area with other registered
areas if possible.

Add QCocoaNativeInterface::testContentBorderPosition(). 
This function tests whether the given point is within 
the unified title and toolbar area.

Use testContentBorderPosition() in QMacStyle to enable 
code paths that skips drawing the QToolBar bottom 
separator line and paints the active tab background 
with transparent pixels to make the background gradient 
visible.

Change-Id: I2b70f9bb0c2c59af053a691a7df538f958783dab
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>