Run non-xcb X tests also with -no-xcb
Passing -no-xcb disables xcb and related configuration tests
completely so for example QT_NO_XRENDER is not defined.  This results
in linker errors in glxconvenience with -no-xcb since the XRender
calls are compiled in without linking to the library.

The XRender, XInput and such tests are not strictly xcb related and
may be used from elsewhere, for example glxconvenience which has
nothing to do with xcb. Therefore these tests are moved in configure
so that they are run even when xcb is disabled.

Task-number: QTBUG-35644
Change-Id: I77871612ea5f6711ecafd8ca53aac6c516c19e2e
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>